backend/api/requirements.txt
Django<4.0
djangorestframework
psycopg2
tzdata
django-cors-headers
django-request-logging
requests
djoser>2.0
djangorestframework_simplejwt
drf_yasg
django-multiselectfield
django-redis